move installer classes to separate files and rename tests
